+++ src/extension.ts
@@ -1,7 +1,10 @@
 import * as vscode from 'vscode';
 import { getNonce, getWebviewContent, handleInsert } from './extension-util';
+import { Field } from './rpg-structure-model';
 
+// Function that activates the extension
 export function activate(context: vscode.ExtensionContext) {
+
 	context.subscriptions.push(
 		vscode.commands.registerCommand('rpg-structure.generator', async () => {
 
@@ -24,6 +27,9 @@ export function activate(context: vscode.ExtensionContext) {
 			// Stores the position to insert code
 			const insertPosition = editor.selection.active;
 
+			// Initialize the field info
+			let fields : Field[] = [];
+
 			// Create Webview
 			const panel = vscode.window.createWebviewPanel(
 				'rpgStructureWizard',
@@ -41,7 +47,7 @@ export function activate(context: vscode.ExtensionContext) {
 				switch (message.command) {
 					// Insert code of the RPG structure
 					case 'insert':
-						handleInsert(editor, insertPosition, message.structureName, message.structureType, message.dimension, message.fields);
+						handleInsert(editor, insertPosition, message.structureName, message.structureType, message.dimension, fields);
 						panel.dispose();
 						break;
 
@@ -49,10 +55,29 @@ export function activate(context: vscode.ExtensionContext) {
 					case 'cancel':
 						panel.dispose();
 						break;
+
+					// Show alert
+					case 'alert':
+						vscode.window.showErrorMessage(message.alert);
+						break;
+
+					// Add new field to structure
+					case 'newfield':
+						
+						// Stores the information of the field
+						const field : Field = {
+							name : message.name,
+							type : message.type,
+							length : message.length,
+							init : message.init
+						};
+						fields.push(field);
+						break;
 				};
 			});
 		})
 	);
 };
 
+// Deactivates the extension
 export function deactivate() {}