+++ src/rpg-structure-model.ts
@@ -1,0 +1,7 @@
+
+export interface Field {
+	name : string;
+	type : string;
+	length : string;
+	init : string;
+};