+++ CHANGELOG.md
@@ -3,8 +3,13 @@
 ## Changed
 v0.0.2
 - Different structure of the project.
+v0.0.3
+- Internal changes.
+- More information in the README.md.
 
 
 ## Added
 v0.0.2
 - The code is inserted from the place it was executed (also if the position is not the first column of the code).
+v0.0.3
+- Added extension to the activity bar.